reorganize module structure
