+++ mock.go
@@ -8,6 +8,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"math/big"
+	. "mergemock/api"
 	"os"
 	"time"
